_archive/requirements.txt
requests
pandas
networkx
haversine
osmnx
folium
numpy
scikit-learn
python-dotenv
streamlit
streamlit-folium
tqdm
watchdog